+++ pyrefly.toml
@@ -232,8 +232,9 @@ project-includes = [
   "nemo_rl/models/policy/workers/__init__.py",
   "nemo_rl/models/policy/workers/megatron_remote_sparse_refit.py",
   "nemo_rl/models/policy/workers/checkpoint_engine.py",
-  "nemo_rl/models/policy/workers/mamba_zero_kl_patches.py",
-  "nemo_rl/models/policy/workers/moe_zero_kl_patches.py",
+  "nemo_rl/models/generation/megatron/zero_train_gen_kl_patches/core_patches.py",
+  "nemo_rl/models/generation/megatron/zero_train_gen_kl_patches/mamba_zero_kl_patches.py",
+  "nemo_rl/models/generation/megatron/zero_train_gen_kl_patches/moe_zero_kl_patches.py",
   "nemo_rl/models/policy/workers/patches.py",
   "nemo_rl/models/value/__init__.py",
   "nemo_rl/models/value/config.py",
